+++ tests/systemtests/suite/query.go
@@ -147,7 +147,7 @@ func (s *SystemTestSuite) cosmosTxPoolContent(nodeID string) (pendingTxHashes, q
 
 	pendingtxHashes := make([]string, 0)
 	for _, tx := range result.Txs {
-		pendingtxHashes = append(pendingtxHashes, tx.Hash().String())
+		pendingtxHashes = append(pendingtxHashes, string(tx.Hash()))
 	}
 
 	return pendingtxHashes, nil, nil